requirement.txt
pyserial
python-dotenv
pyotp
Flask
Flask-Mail
requests
